新增方案自定义图标功能

+++ WeaselServer/WeaselServerApp.h
@@ -15,17 +15,17 @@ class WeaselServerApp {
 public:
 	static bool execute(const std::wstring &cmd, const std::wstring &args)
 	{
-		return (int)ShellExecuteW(NULL, NULL, cmd.c_str(), args.c_str(), NULL, SW_SHOWNORMAL) > 32;
+		return (size_t)ShellExecuteW(NULL, NULL, cmd.c_str(), args.c_str(), NULL, SW_SHOWNORMAL) > 32;
 	}
 
 	static bool explore(const std::wstring &path)
 	{
-		return (int)ShellExecuteW(NULL, L"open", L"explorer", path.data(), NULL, SW_SHOWNORMAL) > 32;
+		return (size_t)ShellExecuteW(NULL, L"open", L"explorer", path.data(), NULL, SW_SHOWNORMAL) > 32;
 	}
 
 	static bool open(const std::wstring &path)
 	{
-		return (int)ShellExecuteW(NULL, L"open", path.c_str(), NULL, NULL, SW_SHOWNORMAL) > 32;
+		return (size_t)ShellExecuteW(NULL, L"open", path.c_str(), NULL, NULL, SW_SHOWNORMAL) > 32;
 	}
 
 	static bool check_update()
